+++ clang-tools-extra/test/clang-tidy/checkers/readability/use-anyofallof.cpp
@@ -1,6 +1,9 @@
 // RUN: %check_clang_tidy -std=c++14,c++17 %s readability-use-anyofallof %t -- -- -fexceptions
 // FIXME: Fix the checker to work in C++20 mode.
 
+#include <utility>
+#include <vector>
+
 bool good_any_of() {
   int v[] = {1, 2, 3};
   // CHECK-MESSAGES: :[[@LINE+1]]:3: warning: replace loop by 'std::any_of()' [readability-use-anyofallof]
@@ -178,6 +181,38 @@ bool good_all_of() {
   int v[] = {1, 2, 3};
   // CHECK-MESSAGES: :[[@LINE+1]]:3: warning: replace loop by 'std::all_of()' [readability-use-anyofallof]
   for (int i : v)
+    if (i)
+      return false;
+  return true;
+}
+
+std::vector<int> get_dummy_vec();
+
+bool good_any_of_temporary_vector() {
+  for (int i : get_dummy_vec())
+    if (i)
+      return true;
+  return false;
+}
+
+bool good_all_of_temporary_vector() {
+  for (int i : get_dummy_vec())
+    if (i)
+      return false;
+  return true;
+}
+
+bool good_xvalue(std::vector<int>& v) {
+  // CHECK-MESSAGES: :[[@LINE+1]]:3: warning: replace loop by 'std::any_of()' [readability-use-anyofallof]
+  for (int i : std::move(v))
+    if (i)
+      return true;
+  return false;
+}
+
+bool good_xvalue_all_of(std::vector<int>& v) {
+  // CHECK-MESSAGES: :[[@LINE+1]]:3: warning: replace loop by 'std::all_of()' [readability-use-anyofallof]
+  for (int i : std::move(v))
     if (i)
       return false;
   return true;